// server/type.ts
export type BigHeadOptions = {
    accessory: string[];
    body: string[];
    circleColor: string[];
    clothing: string[];
    clothingColor: string[];
    eyebrows: string[];
    eyes: string[];
    facialHair: string[];
    graphic: string[];
    hair: string[];
    hairColor: string[];
    hat: string[];
    hatColor: string[];
    lashes: boolean[];
    lipColor: string[];
    mask: boolean[];
    faceMask: boolean[];
    mouth: string[];
    skinTone: string[];
    faceMaskColor: string[];
  };

  